renamed main to start and init LoginFragment
